pyproject.toml
[project]
name = "tsubame-distributed-training-example"
version = "1.0.0"
description = "tsubame: Distributed Training Example on TSUBAME 4.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "mpi4py>=4.1.1",
    "numpy>=2.4.3",
    "rich>=14.3.3",
    "torch>=2.10.0",
    "tyro>=1.0.10",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
]

[tool.uv]
default-groups = ["dev"]
no-build-isolation-package = []

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.setuptools]
packages = ["tsubame"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 99
target-version = "py313"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/linter/#rule-selection
select = ["E", "F", "UP", "B", "SIM", "I", "W", "N", "ANN"]
ignore = [
  "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in ...
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["tsubame"]

[project.scripts]
tsubame-train = "tsubame.scripts.train:entrypoint"
tsubame-launch = "tsubame.scripts.launch:entrypoint"
